redoing adminManagement.js to include bootstrap
